Add surfing and splashy trie (#9)

+++ bench/SuRFFacade.h
@@ -2,15 +2,15 @@
 #define RANGE_FILTERING_SURFFACADE_H
 
 #include <vector>
-#include <surf.hpp>
 #include <PrefixFilter.h>
+#include "../external/SuRF/include/surf.hpp"
 
 namespace range_filtering {
 class SuRFFacade : public PrefixFilter {
 public:
     explicit SuRFFacade(std::vector<std::string> &keys, bool real, uint32_t bits_cnt);
     bool lookupPrefix(const std::string &prefix) override;
-    uint64_t getMemoryUsage() const override;
+    unsigned long long getMemoryUsage() const override;
     std::string getName() const override { return name_; }
 private:
     surf::SuRF *surf_;
@@ -34,7 +34,7 @@ bool SuRFFacade::lookupPrefix(const std::string &prefix) {
     return surf_->lookupRange(prefix, true, right_key, false);
 }
 
-uint64_t SuRFFacade::getMemoryUsage() const {
+unsigned long long SuRFFacade::getMemoryUsage() const {
     return surf_->getMemoryUsage();
 }
 }
