+++ include/BlindTrieWithBloomFilter.h
@@ -15,7 +15,7 @@ public:
     // Keys must be sorted
     explicit BlindTrieWithBloomFilter(std::vector<std::string> &keys, uint32_t bloom_filter_size);
     bool lookupPrefix(const std::string &prefix) override;
-    uint64_t getMemoryUsage() const override;
+    unsigned long long getMemoryUsage() const override;
     std::string getName() const override { return "BloomTrie " + std::to_string(bloomFilter_->getSize()); }
 
 private:
@@ -32,8 +32,8 @@ private:
         void insertChildNode(char current, uint64_t position,
                              std::vector<std::string> &keys,
                              std::vector<std::string> &hashed_substrings);
-        virtual uint64_t getMemoryUsage() const;
-        uint64_t getChildrenMemoryUsage() const;
+        virtual unsigned long long getMemoryUsage() const;
+        unsigned long long getChildrenMemoryUsage() const;
         static void generatePrefixes(std::string &word, uint64_t start_position, std::vector<std::string> &hashed_prefixes);
 
         friend class BlindTrieWithBloomFilter;
@@ -42,7 +42,7 @@ private:
     class BlindNode : public Node {
     protected:
         BlindNode(uint64_t fingerprint, uint64_t length);
-        uint64_t getMemoryUsage() const override;
+        unsigned long long getMemoryUsage() const override;
 
         uint8_t fingerprint_;
         uint64_t length_;
