+++ include/MultiPrefixBloomFilter.h
@@ -18,7 +18,7 @@ public:
                                     MemoryAllocationType memoryAllocationType = MemoryAllocationType::proportional,
                                     double maximalDecreasePercentage = 0.0);
     bool lookupPrefix(const std::string& prefix) override;
-    uint64_t getMemoryUsage() const override;
+    unsigned long long getMemoryUsage() const override;
     std::string getName() const override { return "MultiPrefixBF"; }
 
     std::vector<uint64_t> getBFsSizes();
